+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Not yet released
 
+- Add support for Erlang/OTP20.
 - Improve tracer usage text.
 - Fix `vmq_diversity` memcached issue (#460).
 - Fix RPM package issue preventing a clean upgrade.